Add AuthService, AuthDelegate and force AuthController to implement AuthDelegate.
AuthService is a part of the core application logic, handling
authentication - namely registration and logging in. Depending on the
outcome, the AuthService will instruct the AuthDelegate (in this case,
our controller) to carry out a method. For example, if login is
successful, we tell the AuthDelegate to fire the userLoggedIn() method.